requirements.txt
cffi
cfnresponse
pyOpenSSL
